refactor: move survey's data into the survey folder
