+++ carrerait-roadmap/src/data/surveys/unemployedQuestions.ts
